requirements.txt
requests
fastapi
uvicorn
requests
websocket-client
